requirements.txt
pandas
rdkit-pypi
streamlit
rsconnect-python